+++ rules/eslint.js
@@ -4,18 +4,22 @@ module.exports = {
   ].map(require.resolve),
 
   // View link below for eslint rules documentation
-  // https://eslint.org/docs/latest/rules/
+  // https://eslint.org/docs/rules/
   rules: {
     // The airbnb config enforce that class methods use "this", but requires changing how you call the method
-    // https://eslint.org/docs/latest/rules/class-methods-use-this
+    // https://eslint.org/docs/rules/class-methods-use-this
     "class-methods-use-this": "off",
 
+    // The airbnb config disables the complexity, but we want to limit it
+    // https://eslint.org/docs/rules/complexity
+    complexity: ["error", 20],
+
     // The airbnb config enforces consistent return, but we got problems with some external libraries, TypeScript enforces this anyway
-    // https://eslint.org/docs/latest/rules/consistent-return
+    // https://eslint.org/docs/rules/consistent-return
     "consistent-return": "off",
 
     // The airbnb config forces unix style, but somebody also works on windows
-    // https://eslint.org/docs/latest/rules/linebreak-style
+    // https://eslint.org/docs/rules/linebreak-style
     "linebreak-style": "off",
 
     // The airbnb config limits to 100 ignoring some lines, instad we allow 160 but don't ignore lines
@@ -29,12 +33,20 @@ module.exports = {
       ignoreRegExpLiterals: false,
     }],
 
+    // The airbnb disables the max lines, but we want to limit it
+    // https://eslint.org/docs/rules/max-lines
+    "max-lines": ["error", {
+      max: 300,
+      skipBlankLines: true,
+      skipComments: true
+    }],
+
     // The airbnb config disallow await inside of loops, but it cannot always be avoided
     // https://eslint.org/docs/rules/no-await-in-loop
     "no-await-in-loop": "off",
 
     // The airbnb config disallow use of the continue statement, but it's ok to have it
-    // https://eslint.org/docs/latest/rules/no-continue
+    // https://eslint.org/docs/rules/no-continue
     "no-continue": "off",
 
     // The airbnb config disallow use of unary operators (++ and --), but it's ok to have them
@@ -42,7 +54,7 @@ module.exports = {
     "no-plusplus": "off",
 
     // The airbnb config also restrict ForOfStatement, but we want to use it
-    // https://eslint.org/docs/latest/rules/no-restricted-syntax
+    // https://eslint.org/docs/rules/no-restricted-syntax
     "no-restricted-syntax": [
       "error",
       {
@@ -59,9 +71,5 @@ module.exports = {
         message: "`with` is disallowed in strict mode because it makes code impossible to predict and optimize.",
       },
     ],
-
-    // The airbnb config forces single quote, but we prefer double quote
-    // https://eslint.org/docs/latest/rules/quotes
-    quotes: ["error", "double", { avoidEscape: true }],
   },
 };